refactor: study gorup 및 user study group 패키지 이동
