+++ src/main/java/ggs/srr/api/ApiControllerAdvice.java
@@ -2,8 +2,8 @@ package ggs.srr.api;
 
 import ggs.srr.api.controller.jwt.exception.JwtControllerException;
 import ggs.srr.repository.reservation.exception.FindIdNullException;
-import ggs.srr.service.reservation.exception.EmptyUserIdException;
-import ggs.srr.service.reservation.exception.NoSuchUserException;
+import ggs.srr.service.studygroup.exception.EmptyUserIdException;
+import ggs.srr.service.studygroup.exception.NoSuchUserException;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
