+++ src/main/java/ggs/srr/api/controller/studtygroup/StudyGroupController.java
@@ -3,9 +3,9 @@ package ggs.srr.api.controller.studtygroup;
 import ggs.srr.api.ApiResponse;
 import ggs.srr.api.controller.studtygroup.request.StudyGroupCreateRequest;
 import ggs.srr.api.controller.studtygroup.request.UserAllStudyGroupRequest;
-import ggs.srr.service.reservation.exception.EmptyUserIdException;
-import ggs.srr.service.reservation.studygroup.StudyGroupService;
-import ggs.srr.service.reservation.studygroup.response.UserAllStudyGroupResponse;
+import ggs.srr.service.studygroup.exception.EmptyUserIdException;
+import ggs.srr.service.studygroup.StudyGroupService;
+import ggs.srr.service.studygroup.response.UserAllStudyGroupResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
