+++ src/main/java/ggs/srr/domain/user/FtUser.java
@@ -1,7 +1,7 @@
 package ggs.srr.domain.user;
 
 import ggs.srr.domain.projectuser.ProjectUser;
-import ggs.srr.domain.reservation.userstudygroup.UserStudyGroup;
+import ggs.srr.domain.userstudygroup.UserStudyGroup;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
