+++ src/main/java/ggs/srr/service/studygroup/exception/EmptyUserIdException.java
@@ -1,4 +1,4 @@
-package ggs.srr.service.reservation.exception;
+package ggs.srr.service.studygroup.exception;
 
 public class EmptyUserIdException extends RuntimeException {
     public EmptyUserIdException(String message) {
