+++ src/main/java/ggs/srr/service/studygroup/exception/NoSuchUserException.java
@@ -1,4 +1,4 @@
-package ggs.srr.service.reservation.exception;
+package ggs.srr.service.studygroup.exception;
 
 public class NoSuchUserException extends RuntimeException{
     public NoSuchUserException(String message) {
